refactor(rust): splitter send_handlers.rs et smtp_server.rs
- send_handlers.rs (718) → send_pipeline.rs (408) + send_endpoints.rs (314)
- smtp_server.rs (689) → smtp_server.rs (424) + smtp_server_dir/headers.rs (83) + smtp_server_dir/commands.rs (204)
- Aucun changement de comportement. Chaque fichier ≤ 500 LOC.
